refactor(billing): apply post-review refinements to pricing module

+++ src/features/billing/schemas/billing-schema.ts
@@ -10,7 +10,7 @@ export type CheckoutInput = z.infer<typeof checkoutSchema>
 export const pricingPlanCreateSchema = z.object({
   name: z.string().min(1).max(100),
   description: z.string().max(500).optional(),
-  price: z.coerce.number().int().nonnegative(),
+  price: z.coerce.number().nonnegative(),
   features: z.record(z.string(), z.boolean()).optional(),
   stripePriceId: z.string().max(255).optional(),
   isActive: z.boolean().default(true),
